+++ src/datasources/asset/data-sources/calibration-forecast/CalibrationForecastDataSource.ts
@@ -1,13 +1,19 @@
-import { DataQueryRequest, DataFrameDTO, DataSourceInstanceSettings, FieldDTO, TestDataSourceResponse } from '@grafana/data';
-import { AssetDataSourceOptions, AssetQuery, AssetQueryType } from '../../types/types';
+import { DataQueryRequest, DataFrameDTO, DataSourceInstanceSettings, FieldDTO, TestDataSourceResponse, DataLink } from '@grafana/data';
+import { AssetDataSourceOptions, AssetQuery, AssetQueryType, AssetType, AssetTypeOptions, BusType, BusTypeOptions } from '../../types/types';
 import { BackendSrv, getBackendSrv, getTemplateSrv, TemplateSrv } from '@grafana/runtime';
 import { AssetDataSourceBase } from '../AssetDataSourceBase';
-import { AssetCalibrationForecastKey, AssetCalibrationTimeBasedGroupByType, AssetType, AssetTypeOptions, BusType, BusTypeOptions, CalibrationForecastQuery, CalibrationForecastResponse, ColumnDescriptorType, FieldDTOWithDescriptor } from '../../types/CalibrationForecastQuery.types';
+import { AssetCalibrationForecastKey, AssetCalibrationTimeBasedGroupByType, CalibrationForecastQuery, CalibrationForecastResponse, ColumnDescriptorType, FieldDTOWithDescriptor } from '../../types/CalibrationForecastQuery.types';
 import { transformComputedFieldsQuery } from '../../../../core/query-builder.utils';
 import { AssetModel, AssetsResponse } from '../../../asset-common/types';
 
 export class CalibrationForecastDataSource extends AssetDataSourceBase {
     private dependenciesLoadedPromise: Promise<void>;
+
+    private readonly forecastDateFormatterMap = new Map<AssetCalibrationForecastKey, (date: string) => string>([
+        [AssetCalibrationForecastKey.Day, this.formatDateForDay],
+        [AssetCalibrationForecastKey.Week, this.formatDateForWeek],
+        [AssetCalibrationForecastKey.Month, this.formatDateForMonth]
+    ]);
 
     constructor(
         readonly instanceSettings: DataSourceInstanceSettings<AssetDataSourceOptions>,
@@ -46,6 +52,7 @@ export class CalibrationForecastDataSource extends AssetDataSourceBase {
             );
         }
 
+        calibrationForecastQuery.IncludeOnlyDataInTimeRange = this.getQueryParam('IncludeOnlyDataInTimeRange')?.toLowerCase() === 'true';
         return await this.processCalibrationForecastQuery(calibrationForecastQuery, options);
     }
 
@@ -63,11 +70,19 @@ export class CalibrationForecastDataSource extends AssetDataSourceBase {
         const from = options.range!.from.toISOString();
         const to = options.range!.to.toISOString();
 
-        const calibrationForecastResponse: CalibrationForecastResponse = await this.queryCalibrationForecast(query.groupBy, from, to, query.filter);
+        const calibrationForecastResponse: CalibrationForecastResponse = await this.queryCalibrationForecast(
+            query.groupBy,
+            from,
+            to,
+            query.filter,
+            query.IncludeOnlyDataInTimeRange
+        );
 
         result.fields = calibrationForecastResponse.calibrationForecast.columns || [];
-        if (this.isGroupByTime(query)) {
-            this.processResultsGroupedByTime(result)
+
+        const timeGrouping = this.getTimeGroup(query);
+        if (timeGrouping) {
+            this.processResultsGroupedByTime(result, timeGrouping)
         } else {
             this.processResultsGroupedByProperties(result)
         }
@@ -75,29 +90,82 @@ export class CalibrationForecastDataSource extends AssetDataSourceBase {
         return result;
     }
 
-    isGroupByTime(query: CalibrationForecastQuery) {
-        return query.groupBy.includes(AssetCalibrationTimeBasedGroupByType.Day) ||
-            query.groupBy.includes(AssetCalibrationTimeBasedGroupByType.Week) ||
-            query.groupBy.includes(AssetCalibrationTimeBasedGroupByType.Month);
+    getTimeGroup(query: CalibrationForecastQuery) {
+        if (query.groupBy.includes(AssetCalibrationTimeBasedGroupByType.Day)) {
+            return AssetCalibrationForecastKey.Day;
+        }
+        if (query.groupBy.includes(AssetCalibrationTimeBasedGroupByType.Week)) {
+            return AssetCalibrationForecastKey.Week;
+        }
+        if (query.groupBy.includes(AssetCalibrationTimeBasedGroupByType.Month)) {
+            return AssetCalibrationForecastKey.Month;
+        }
+        return null;
     }
 
-    processResultsGroupedByTime(result: DataFrameDTO) {
+    processResultsGroupedByTime(result: DataFrameDTO, timeGrouping: AssetCalibrationForecastKey) {
+        const timeFieldFirstValue = result.fields.find(field => field.name === timeGrouping)?.values?.at(0);
+        const startDate = this.forecastDateFormatterMap.get(timeGrouping)!(timeFieldFirstValue);
+
         result.fields.forEach(field => {
             field.name = this.createColumnNameFromDescriptor(field as FieldDTOWithDescriptor);
-            switch (field.name) {
-                case AssetCalibrationForecastKey.Day:
-                    field.values = field.values!.map(this.formatDateForDay)
-                    break;
-                case AssetCalibrationForecastKey.Week:
-                    field.values = field.values!.map(this.formatDateForWeek)
-                    break;
-                case AssetCalibrationForecastKey.Month:
-                    field.values = field.values!.map(this.formatDateForMonth)
-                    break;
-                default:
-                    break;
+            const formatter = this.forecastDateFormatterMap.get(field.name as AssetCalibrationForecastKey);
+
+            if (formatter) {
+                field.values = field.values!.map(formatter);
+            }
+            
+            if (!formatter && !Boolean(this.getQueryParam('IncludeOnlyDataInTimeRange'))) {
+                field.config = { links: this.createDataLinks(timeGrouping, startDate) };
             }
         });
+    }
+
+    private createDataLinks(timeGrouping: AssetCalibrationForecastKey, startDate: string): DataLink[] {
+        const url = '/d/${__dashboard.uid}/${__dashboard}?orgId=${__org.id}&${__all_variables}';
+
+        return [
+            {
+                title: `View ${timeGrouping}`, targetBlank: true, url: `${url}`,
+                onBuildUrl: function (options) {
+                    if (!options.replaceVariables) {
+                        return url;
+                    }
+
+                    const value = options.replaceVariables(`\${__data.fields.${timeGrouping}}`);
+                    const IncludeOnlyDataInTimeRange = value !== startDate;
+
+                    let from, to;
+
+                    const parseDate = (dateStr: string) => {
+                        const date = new Date(dateStr);
+                        date.setUTCHours(0);
+                        return date;
+                    };
+
+                    switch (timeGrouping) {
+                        case AssetCalibrationForecastKey.Day:
+                            const dayDate = parseDate(value);
+                            from = dayDate.valueOf();
+                            dayDate.setUTCDate(dayDate.getUTCDate() + 1);
+                            to = dayDate.valueOf();
+                            break;
+                        case AssetCalibrationForecastKey.Week:
+                            [from, to] = value.split(' : ').map(dateStr => parseDate(dateStr).valueOf());
+                            break;
+                        case AssetCalibrationForecastKey.Month:
+                            const monthDate = parseDate(value);
+                            monthDate.setUTCDate(monthDate.getUTCDate() + 1);
+                            from = monthDate.valueOf();
+                            monthDate.setUTCMonth(monthDate.getUTCMonth() + 1);
+                            to = monthDate.valueOf();
+                            break;
+                    }
+
+                    return `${url}&from=${from}&to=${to}&IncludeOnlyDataInTimeRange=${IncludeOnlyDataInTimeRange}`;
+                }
+            }
+        ];
     }
 
     processResultsGroupedByProperties(result: DataFrameDTO) {
@@ -137,9 +205,11 @@ export class CalibrationForecastDataSource extends AssetDataSourceBase {
     }
 
     formatDateForWeek(date: string): string {
+        console.log(date);
         const startDate = new Date(date);
         const endDate = new Date(startDate);
         endDate.setDate(startDate.getDate() + 6);
+        console.log(startDate)
         return `${startDate.toISOString().split('T')[0]} : ${endDate.toISOString().split('T')[0]}`;
     }
 
@@ -157,8 +227,8 @@ export class CalibrationForecastDataSource extends AssetDataSourceBase {
         }
     }
 
-    async queryCalibrationForecast(groupBy: string[], startTime: string, endTime: string, filter = ''): Promise<CalibrationForecastResponse> {
-        let data = { groupBy, startTime, endTime, filter };
+    async queryCalibrationForecast(groupBy: string[], startTime: string, endTime: string, filter = '', IncludeOnlyDataInTimeRange = false): Promise<CalibrationForecastResponse> {
+        let data = { groupBy, startTime, endTime, filter, IncludeOnlyDataInTimeRange };
         try {
             let response = await this.post<CalibrationForecastResponse>(this.baseUrl + '/assets/calibration-forecast', data);
             return response;