+++ src/datasources/system/components/SystemQueryEditor.tsx
@@ -75,7 +75,7 @@ export function SystemQueryEditor({ query, onChange, onRunQuery, datasource }: P
 }
 
 const tooltips = {
-  queryType: `Properties allows you to visualize the properties of one or more systems.
+  queryType: `Properties allows you to visualize one or more systems' properties.
               Summary allows you to visualize the number of disconnected and connected systems.`,
   system: `Query for a specific system by its name or ID. If left blank, the plugin returns all
             available systems. You can enter a variable into this field.`,